move computation from tauri-commands to backend
